Moved ResultSelectionService global services to ResultSelectionModule
