+++ frontend/src/app/modules/application-dashboard/components/performance-aspect-management/performance-aspect-inspect/performance-aspect-inspect.component.spec.ts
@@ -4,7 +4,7 @@ import {PerformanceAspectInspectComponent} from './performance-aspect-inspect.co
 import {SharedMocksModule} from "../../../../../testing/shared-mocks.module";
 import {MeasurandSelectComponent} from "../../../../result-selection/components/measurand-select/measurand-select.component";
 import {ApplicationService} from "../../../../../services/application.service";
-import {ResultSelectionService} from "../../../../../services/result-selection.service";
+import {ResultSelectionService} from "../../../../result-selection/services/result-selection.service";
 
 describe('PerformanceAspectInspectComponent', () => {
   let component: PerformanceAspectInspectComponent;
