+++ frontend/src/app/modules/result-selection/components/measurand-select/measurand-select.component.spec.ts
@@ -2,6 +2,7 @@ import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 
 import {MeasurandSelectComponent} from './measurand-select.component';
 import {SharedMocksModule} from "../../../../testing/shared-mocks.module";
+import {ResultSelectionService} from "../../services/result-selection.service";
 
 describe('MeasurandSelectComponent', () => {
   let component: MeasurandSelectComponent;
@@ -10,7 +11,10 @@ describe('MeasurandSelectComponent', () => {
   beforeEach(async(() => {
     TestBed.configureTestingModule({
       declarations: [MeasurandSelectComponent],
-      imports: [SharedMocksModule]
+      imports: [SharedMocksModule],
+      providers: [
+        ResultSelectionService
+      ]
     })
     .compileComponents();
   }));
