+++ frontend/src/app/modules/result-selection/result-selection.module.ts
@@ -2,6 +2,7 @@ import {NgModule} from '@angular/core';
 import {MeasurandSelectComponent} from "./components/measurand-select/measurand-select.component";
 import {SharedModule} from "../shared/shared.module";
 import {FormsModule} from "@angular/forms";
+import {ResultSelectionService} from "./services/result-selection.service";
 
 @NgModule({
   imports: [
@@ -9,6 +10,9 @@ import {FormsModule} from "@angular/forms";
     FormsModule
   ],
   declarations: [MeasurandSelectComponent],
-  exports: [MeasurandSelectComponent]
+  exports: [MeasurandSelectComponent],
+  providers: [
+    ResultSelectionService
+  ]
 })
 export class ResultSelectionModule { }
